Add keyword in plot_airmass for Moon Airmass, brightness_shading multiple days
I couldn't find a way to do this with the existing astroplan system. Please correct me if I'm wrong and missed something, It seems like something people would want to be able to add on to their plot so I figured I should add it.
Also, I added in a quick fix to show brightness shading more than one day i.e. if I want to see it over a timeframe of several days, which it didn't do before.