convert quantities before they might be strings
trying to `u.Quantity` with a string, for example, `u.Quantity([m42.ra, ''])`,
yields a `TypeError`:
```
    print(u.Quantity([m42.ra, '']))
  File "~/.local/lib/python3.8/site-packages/astropy/units/quantity.py", line 382, in __new__
    raise TypeError("The value must be a valid Python or "
TypeError: The value must be a valid Python or Numpy numeric type.
```

this is exactly what happens in `to_table()` with transitions or unused slots
shown, because of the append `append('')` in `scheduling.py`.

so instead, do the `u.Quantity` conversion earlier where we know the type is
(should be!) a Numpy numeric.